Move gammapy/modeling/models/spectrum/utils.py -> gammapy/utils/integrate.py
