+++ gammapy/utils/serialization/xml.py
@@ -11,8 +11,7 @@ import numpy as np
 import astropy.units as u
 from gammapy.extern import xmltodict
 from gammapy.maps import Map
-from gammapy.modeling.models import SkyModel, SkyModels
-from gammapy.modeling.models import spatial
+from gammapy.modeling.models import SkyModel, SkyModels, spatial
 from gammapy.modeling.models import spectrum as spectral
 from gammapy.utils.fitting import Parameter, Parameters
